backend/pyproject.toml
[project]
name = "ai-journal"
version = "0.0.1"
description = ""
authors = [
    {name = "Luan van Pletsen", email = "luanvanpletsen@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "openai>=1.16.2",
    "ipykernel>=6.29.4",
    "fastapi>=0.110.1",
    "uvicorn>=0.29.0",
    "pre-commit>=3.7.0",
    "weaviate-client==4.6.3",
    "rag-router",
    "dspy-ai[weaviate]>=2.5.12",
    "pip>=24.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_journal"]

[tool.uv]
dev-dependencies = [
    "pytest-mock>=3.14.0",
    "pytest>=8.3.3",
]

[tool.uv.sources]
rag-router = { git = "https://github.com/Luan-vP/rag_router", rev = "fc5a73c67cc721bd7b7fad91a24e741777cd4ba3" }
